Added stock code to make API request to Plantnet
The code added is not formatted for this project, so I will need to adjust the parameters as well as the workflow for the API Request to function. Still need to connect the user submitted photo to the API request, and then connect the results with a display and additional information.