Slight rework in progress
I didn't realize that the UI protocol that I was using in the code in place program was proprietary to that course, and so I had written a basic UI entirely there and cannot utilize the same work when trying to deploy out here in the real world. I'm considering which platform would be best to utilize, whether it's Tkinter, EasyGUI, or even PyScript. PyScript seems the most interesting to me because I wouldn't need a backend server to host the final product. The entire project could operate from any web browser, even ones on mobile phones if I'm not mistaken. This would be a huge workaround for the backend issue and so if PyScript is indeed the answer I'm looking for, then I will be focusing on retooling the UI with webdev tools in order to utilize PyScript. This commit has the main program a bit sloppily labeled with areas meant to be retooled. I've also added the functionality to prompt the user to select a file path from their local drive. This will be the initial feature in the program to utilize local files to submit to the PlantNet API. The Main.html file contains the beginnings of the HTML UI to be hosted on PyScript.