A new path and a new endgame
After researching a bit more, it looks like almost a complete rework of the codebase is necessary to format the program to work with Pyscript. Pyscript still feels like the best venue for this program until I can get into backend work or develop the mobile app codebase. In this rework I'll need to create the UI using web dev tools, and the HTML pages will host the Python scripts for the functionality of the program.

This commit includes new files "Python_1_index.py" is a revised page of "plant_indentifier.py" where now it is just the functionality of prompting the user for a filepath, or series of filepaths.

"Index.html" and "style.css" are the beginnings of the web buildout for PyScript.