Rework continues
Moved older ignored files to a legacy folder and am continuing to reformat the program to fit the needs of the new deployment plan: Build the GUI with web dev tools, and then call in python functions to send the API request. Continued to refine the HTML and build out the different pages necessary for the UI. Created JSON doc.