Almost there - Final form taking shape
This commit introduces the "results" page where the uploaded photos, and the returned results from the API will be displayed. Debugging has begun, although there doesn't appear to be too many issues, if the co-pilot in VSC is to be believed. Smoothed out the API request language, and made sure that the appropriate modules are called in.