AI enters the chat, and the case for starting over
I followed a friend's recommendation to begin working utilizing AI agents in my workflow, as the technology is apparently here to stay, and is how people will remain competitive in this industry. So today I started working with Github's CoPilot on debugging the program. It turns out that the interworking between Javascript and Python, and then also with PyScript is rather complex and feels advanced way beyond my current skillset. The agent was able to debug nearly all the issues the program was facing, and now it's formatting the API request that is the focus of the work.

However, I trusted my instinct when I asked the agent if this program was becoming unnecessarily complicated and if the code could be more streamlined, simple, and easier to read if I committed to just using javascript for this project. The agent agreed that since the program currently didn't utilize any python-specific functionality like machine learning or data manipulation, that it was a good idea to make the switch. Coincidentally, PyScript becomes irrelevant considering that javascript can operate entirely in a web browser so there's no need for a system back-end.

I suppose learning that there's a another route to take, one that is easier to navigate, is a pleasant surprise. The fact that I will have to essentially return to square one with this project is an unpleasant surprise. But I suppose this is the learning journey.

All legacy files have been moved to the legacy folder. The next commit will begin the transitioning into a full javascript codebase.